requirements.txt
openai
pandas
tqdm
python-dotenv
wandb
seaborn
fastapi
pydantic
chainlit
markdown2
weasyprint